+++ src/graph.js
@@ -84,12 +84,11 @@ Graph.prototype = {
 
     this.compute(this.args.nodes, this.args.links);
     this.computeScales();
-    
     this.initZoom();
 
     el.html("");
 
-    var svg = el.attr("class", el.attr("class") + " " + BASE_ELEMENT_CLASS) 
+    var svg = el.attr("class", BASE_ELEMENT_CLASS) 
       .append("svg")
         .attr("width", this.opts.width)
         .attr("height", this.opts.height)
@@ -97,30 +96,25 @@ Graph.prototype = {
         .call(this._zoom.on("zoom", bind(this, this.onZoom))
                         .scaleExtent(this.opts.zoomScaleExtent));
 
-    //var markerWidth = 6,
-    //    markerHeight = 6,
-    //    cRadius = 30, // play with the cRadius value
-    //    refX = cRadius + (markerWidth * 2),
-    //    refY = -Math.sqrt(cRadius),
-    //    drSub = cRadius + refY;
-
-    //// Per-type markers, as they don't inherit styles.
-    //svg.append("svg:defs").selectAll("marker")
-    //    .data(["arrow"])
-    //    .enter().append("svg:marker")
-    //    .attr("id", String)
-    //    .attr("viewBox", "0 -5 10 10")
-    //    .attr("refX", refX)
-    //    .attr("refY", refY)
-    //    .attr("markerWidth", markerWidth)
-    //    .attr("markerHeight", markerHeight)
-    //    .attr("orient", "auto")
-    //    .append("svg:path")
-    //    .attr("d", "M0,-5L10,0L0,5");
-
     this.parent = svg.append('svg:g').style('display','none');
 
-    el.on("click", function() { self.reset() });
+    var clickedX, clickedY;
+
+    el.on("mousedown", function() { 
+      clickedX = d3.event.x; 
+      clickedY = d3.event.y; 
+    });
+
+    el.on("click", function() { 
+
+        // avoid reset when dragging
+        if (d3.event.x != clickedX || d3.event.y != clickedY) { 
+          return;
+        }
+        
+        self.reset() 
+      });
+
   },
 
   /**
@@ -130,8 +124,11 @@ Graph.prototype = {
    */
 
   initZoom: function() {
-    this._zoom = d3.behavior.zoom()
-      .center([this.opts.width/2, this.opts.height/2]);
+    this._zoom = d3.behavior.zoom();
+
+    if (this.opts.initialScale) {
+      this._zoom = this._zoom.scale(this.opts.initialScale);
+    }
   },
 
   /**
@@ -341,12 +338,6 @@ Graph.prototype = {
     return this;
   },
 
-  getCurrentCenter: function() {
-    var point = [this.opts.width/2, this.opts.height/2];
-
-    return this.location(point);
-  },
-
   /**
    * Applies zoom in
    *
@@ -378,7 +369,7 @@ Graph.prototype = {
   },
 
   /**
-   * Given, the current zoom and size of a node, tells if it's label should
+   * Given the current zoom and size of a node, tells if it's label should
    * be displayed
    *
    * @api private
@@ -433,10 +424,6 @@ Graph.prototype = {
 
     this.init();
     
-    if (this.opts.initialScale) {
-      this._zoom = this._zoom.scale(this.opts.initialScale);
-    }
-
     function circleRadius(d) { return self.radiusScale(self.getSize(d) || 1); }
 
     var force = this.force = d3.layout.force()
@@ -455,7 +442,6 @@ Graph.prototype = {
       .attr("stroke", bind(this, this.pathStroke))
       .attr("stroke-width", PATH_STROKE_WIDTH)
       .attr("fill", "none");
-      //.attr("marker-end", "url(#arrow)");
     
     var node = this.d3Nodes = this.parent.selectAll(".node")
       .data(force.nodes())
@@ -488,7 +474,7 @@ Graph.prototype = {
         self.updateLinksPosition();
 
         // center the graph
-        self.center();
+        self._center();
 
         // showing canvas after finished rendering
         self.show();
@@ -881,7 +867,6 @@ Graph.prototype = {
       this._focus(focusFn);
     }
 
-
     // store the currently applied filters
     this.appliedFilters = this.filters;
 
@@ -892,13 +877,13 @@ Graph.prototype = {
     this.updateTitles();
 
     // flushing filters
-    
     this.filters = [];
 
     // pending center
     if (this.unappliedCenter) {
       this.unappliedCenter = false;
-      this.center();
+      this._center();
+      this.refreshZoom(true);
     }
 
     // With the currently applied filters we found no matching nodes
@@ -1165,12 +1150,11 @@ Graph.prototype = {
         throw new Error('invalid argument');
     }
 
-    // XXX : only do this when a fn is passed
-    //n = this.findFocusedNode(fn);
-
     if (n) {
       this.setFocus(n);
     }
+
+    this.unappliedFocus = null;
 
     return this;
   },
@@ -1367,7 +1351,6 @@ Graph.prototype = {
     }
 
     this.translateTo([this.opts.width/2, this.opts.height/2], l);
-    //this.refreshZoom(true);
   },
 
   /**