+++ History.md
@@ -3,6 +3,7 @@
 * Improved `zoom()` centering.
 * Fixed bug: `center()` return value.
 * Fixed bug: apply center if called before rendering.
+* Fixed bug: avoid reset when dragging.
 
 # 0.11
 * Added `focus()` with incoming/outgoing relations.